backend/python/requirements.txt
fastapi
uvicorn
pydantic
torch
demucs
python-multipart
python-jose[cryptography]
httpx
